refactor: 코드리뷰 반영
- README.md 통합
- controller 역할이 더 드러나도록 game 클래스 수정
- 클래스 수정에 따른 tc 대응
