+++ README.md
@@ -7,3 +7,14 @@
 
 ## 온라인 코드 리뷰 과정
 * [텍스트와 이미지로 살펴보는 온라인 코드 리뷰 과정](https://github.com/next-step/nextstep-docs/tree/master/codereview)
+
+## 구현목록
+- [ ] 경주 조건을 입력받는다
+- [ ] 게임을 만든다
+- [ ] 게임을 시작한다
+- [ ] 주어진 횟수만큼의 이동을 시도한다
+- [ ] 전진 조건을 판단한다
+- [ ] 전진 조건을 만족하면 진한다
+- [ ] 멈춘다
+- [ ] 경주 결과를 출력한다
+
