--- src/main/java/carracing/README.md
@@ -1,12 +1,0 @@
-# Step3 자동차 경주 
-
-## TODO  
-- [ ] 경주 조건을 입력받는다
-- [ ] 게임을 만든다 
-- [ ] 게임을 시작한다
-- [ ] 주어진 횟수만큼의 이동을 시도한다
-- [ ] 전진 조건을 판단한다
-- [ ] 전진 조건을 만족하면 진한다 
-- [ ] 멈춘다
-- [ ] 경주 결과를 출력한다
-
