+++ src/main/java/carracing/SimpleCarRacing.java
@@ -1,7 +1,6 @@
 package carracing;
 
-import carracing.model.Cars;
-import carracing.controller.Game;
+import carracing.controller.GameController;
 import carracing.view.ConsoleGameDashBoard;
 import carracing.view.ConsoleInput;
 import carracing.view.RacingCondition;
@@ -11,10 +10,8 @@ public class SimpleCarRacing {
     public static void main(String[] args) {
         RacingCondition racingCondition = new ConsoleInput().getInputs();
 
-        Cars cars = new Cars(racingCondition.carCount());
+        GameController gameController = new GameController(new ConsoleGameDashBoard());
 
-        Game game = new Game(racingCondition.attemptCount(), cars, new ConsoleGameDashBoard());
-
-        game.play();
+        gameController.play(racingCondition);
     }
 }
