+++ src/test/java/carracing/model/GameRuleTest.java
@@ -1,10 +1,8 @@
-package carracing.controller;
+package carracing.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import carracing.model.GameRule;
 
 public class GameRuleTest {
